go v4 style for workflow (#1197)
* go v4 style for workflow

* remove some todos
